feat and refactor: connect catalog and details, refactor logic of detail page
