--- src/features/product/product-info/index.ts
@@ -1,5 +1,0 @@
-import productInfoController from "./model/productInfoController";
-
-const productInfo = productInfoController.getView();
-
-export default productInfo;
