--- src/features/product/product-info/model/productInfoModel.ts
@@ -1,5 +1,0 @@
-export class ProductInfoModel {}
-
-const productInfoModel = new ProductInfoModel();
-
-export default productInfoModel;
