+++ src/pages/product-details/api/getProductData.ts
@@ -1,4 +1,4 @@
-import requestAPI from "../../../../shared/api/APIRootBuilder";
+import requestAPI from "../../../shared/api/APIRootBuilder";
 
 export default function getProductData(key: string) {
   return requestAPI
