--- src/pages/product-details/api/index.ts
@@ -1,6 +1,0 @@
-import requestAPI from "../../../shared/api/APIRootBuilder";
-
-const fetchProductByProductKey = (key: string) =>
-  requestAPI.apiRoot().productProjections().withKey({ key }).get().execute();
-
-export default fetchProductByProductKey;
