--- src/pages/product-details/ui/productDetails.scss
@@ -1,67 +1,0 @@
-.product-details {
-  overflow: hidden;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  max-width: 380px;
-  background: var(--additional-pages-background);
-  border-radius: 10px;
-  // cursor: pointer;
-  // transition: 0.3s;
-
-  // &:hover {
-  //   transform: scale(1.005);
-  // }
-
-  // &:active {
-  //   transform: scale(1);
-  // }
-}
-.product-details__img {
-  max-width: 100%;
-  max-height: 100%;
-  transition: 0.3s;
-}
-.product-details__content {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-  align-items: center;
-  padding: 1rem;
-}
-.product-details__title {
-  font-size: 1.6rem;
-  font-weight: 500;
-}
-.product-details__description {
-  font-size: 1.2rem;
-  font-weight: 300;
-  line-height: 1.6rem;
-}
-
-.product-details__price {
-}
-.product-details__discount-price {
-}
-.buy-button {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding: 0.5rem 2rem;
-  font-size: 1.2rem;
-  font-weight: 300;
-  color: var(--beige-text-color);
-  background: var(--green-text-color);
-  border: none;
-  border-radius: 10px;
-  cursor: pointer;
-  transition: 0.3s;
-
-  &:hover {
-    transform: scale(1.05);
-  }
-
-  &:active {
-    transform: scale(1);
-  }
-}
